+++ CI/SCRIPTS/test_xml.py
@@ -125,3 +125,46 @@ def test_xml():
     ci.assert_xml_equal(
         xml.read_archive(l8_archived, r".*_MTL\.xml"), xml.read_archive(xml_archived)
     )
+
+
+def test_dict_xml():
+    xml_dict = {
+        "ACCODE": "LaSRC",
+        "AREA_OR_POINT": "Area",
+        "arop_ave_xshift(meters)": "0",
+        "arop_ave_yshift(meters)": "0",
+        "arop_ncp": "0",
+        "arop_rmse(meters)": "0",
+        "arop_s2_refimg": "NONE",
+        "cloud_coverage": "46",
+        "DATASTRIP_ID": "S2B_OPER_MSI_L1C_DS_2BPS_20221127T223847_S20221127T214906_N04.00",
+        "HLS_PROCESSING_TIME": "2022-11-29T07:10:51Z",
+        "HORIZONTAL_CS_CODE": "EPSG:32760",
+        "HORIZONTAL_CS_NAME": "WGS84 / UTM zone 60S",
+        "L1C_IMAGE_QUALITY": "NONE",
+        "L1_PROCESSING_TIME": "2022-11-27T23:01:25.915427Z",
+        "MEAN_SUN_AZIMUTH_ANGLE": "50.8114320532222",
+        "MEAN_SUN_ZENITH_ANGLE": "37.2612285279726",
+        "MEAN_VIEW_AZIMUTH_ANGLE": "104.448494899629",
+        "MEAN_VIEW_ZENITH_ANGLE": "3.7539313224889",
+        "MSI band 01 bandpass adjustment slope and offset": "0.995900,-0.000200",
+        "NBAR_SOLAR_ZENITH": "38.4289484166757",
+        "NCOLS": "3660",
+        "NROWS": "3660",
+        "OVR_RESAMPLING_ALG": "NEAREST",
+        "PROCESSING_BASELINE": "04.00",
+        "PRODUCT_URI": "S2B_MSIL1C_20221127T214909_N0400_R143_T60FXK_20221127T223847.SAFE",
+        "SENSING_TIME": "2022-11-27T21:49:48.678784Z",
+        "SPACECRAFT_NAME": "Sentinel-2B",
+        "spatial_coverage": "53",
+        "SPATIAL_RESOLUTION": "30",
+        "TILE_ID": "S2B_OPER_MSI_L1C_TL_2BPS_20221127T223847_A029913_T60FXK_N04.00",
+        "ULX": "600000",
+        "ULY": "-5499960",
+    }
+
+    root = xml.dict_to_xml(xml_dict)
+
+    for key, val in xml_dict.items():
+        key_xml = key.replace(" ", "_").replace("(", "_").replace(")", "")
+        _assert_str(root.findtext(f".//{key_xml}"), val)